+++ test/topology_custom/mv/tablets/test_mv_tablets.py
@@ -12,7 +12,7 @@ from test.pylib.util import wait_for_cql_and_get_hosts
 from test.pylib.internal_types import ServerInfo
 from test.topology.conftest import skip_mode
 
-from .test_alternator import get_alternator, alternator_config, full_query
+from test.topology_custom.test_alternator import get_alternator, alternator_config, full_query
 
 import pytest
 import asyncio
