+++ test/topology_custom/mv/test_mv_admission_control.py
@@ -12,7 +12,7 @@ import logging
 
 from test.topology.conftest import skip_mode
 from test.pylib.util import wait_for_view
-from test.topology_custom.test_mv_tablets import pin_the_only_tablet, get_tablet_replicas
+from test.topology_custom.mv.tablets.test_mv_tablets import pin_the_only_tablet, get_tablet_replicas
 
 from cassandra.cluster import ConsistencyLevel, EXEC_PROFILE_DEFAULT # type: ignore
 from cassandra.cqltypes import Int32Type # type: ignore
