+++ test/topology_custom/mv/test_mv_delete_partitions.py
